refactor: добавить логирование при изменении объекта
