+++ src/main/java/ru/yandex/practicum/filmorate/models/AppData.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.model;
+package ru.yandex.practicum.filmorate.models;
 
 public abstract class AppData {
     protected int id;
