+++ src/main/java/ru/yandex/practicum/filmorate/models/Film.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.model;
+package ru.yandex.practicum.filmorate.models;
 
 import lombok.Data;
 import org.springframework.format.annotation.DateTimeFormat;
